add eigen, highs as dependencies
Adds Highs as a CMake dependency in the main CPP CMake file.

Add Eigen as a vendored dependency in the shortest_paths module.
Eigen is vendored instead of an alterative due to needing functionality not present in the most recent release.

Add partial implementation and partial tests for polyedge / edge network.